+++ TrelloDotNet/TrelloDotNet/AutomationEngine/Model/Actions/AddChecklistToCardAction.cs
@@ -1,6 +1,9 @@
-﻿using System.Linq;
+﻿using System.Collections.Generic;
+using System.Linq;
+using System.Threading;
 using System.Threading.Tasks;
 using TrelloDotNet.AutomationEngine.Interface;
+using TrelloDotNet.Control;
 using TrelloDotNet.Model;
 using TrelloDotNet.Model.Webhook;
 
@@ -18,6 +21,11 @@ namespace TrelloDotNet.AutomationEngine.Model.Actions
         /// The Checklist Object to Add (if a checklist with same name is not already present)
         /// </summary>
         public Checklist ChecklistToAdd { get; set; }
+
+        /// <summary>
+        /// By default a checklist is only added if it do not already exist. This determine if it already exist if the check-items should be added to the existing list or not
+        /// </summary>
+        public bool AddCheckItemsToExistingChecklist { get; set; }
 
         /// <summary>
         /// Constructor
@@ -51,8 +59,33 @@ namespace TrelloDotNet.AutomationEngine.Model.Actions
             }
             else
             {
-                processingResult.AddToLog($"SKIPPED: Checklist '{ChecklistToAdd.Name}' was already on card '{webhookAction.Data.Card.Name}'");
-                processingResult.ActionsSkipped++;
+                if (AddCheckItemsToExistingChecklist)
+                {
+                    var checklistItemsMissing = ChecklistToAdd.Items.Where(checklistItem => existing.Items.All(x => x.Name != checklistItem.Name)).ToList();
+                    if (checklistItemsMissing.Any())
+                    {
+                        decimal maxPosition = existing.Items.Max(x => x.Position);
+                        foreach (var checklistItem in checklistItemsMissing)
+                        {
+                            maxPosition += 1;
+                            checklistItem.Position = maxPosition;
+                        }
+
+                        await webhookAction.TrelloClient.AddCheckItemsAsync(existing, checklistItemsMissing.ToArray());
+                        processingResult.AddToLog($"Added '{checklistItemsMissing.Count}' items to checklist '{existing.Name}' for card '{webhookAction.Data.Card.Name}'");
+                        processingResult.ActionsExecuted++;
+                    }
+                    else
+                    {
+                        processingResult.AddToLog($"SKIPPED: Checklist '{ChecklistToAdd.Name}' was already on card '{webhookAction.Data.Card.Name}' and no items missing");
+                        processingResult.ActionsSkipped++;
+                    }
+                }
+                else
+                {
+                    processingResult.AddToLog($"SKIPPED: Checklist '{ChecklistToAdd.Name}' was already on card '{webhookAction.Data.Card.Name}'");
+                    processingResult.ActionsSkipped++;
+                }
             }
         }
     }